+++ selfdrive/controls/lib/lane_planner.py
@@ -1,10 +1,11 @@
 import numpy as np
 from cereal import log
 from common.filter_simple import FirstOrderFilter
-from common.numpy_fast import interp, clip
+from common.numpy_fast import interp, clip, mean
 from common.realtime import DT_MDL
 from common.realtime import sec_since_boot
 from selfdrive.config import Conversions as CV
+from selfdrive.controls.lib.drive_helpers import apply_deadzone
 from selfdrive.hardware import EON, TICI
 from selfdrive.swaglog import cloudlog
 from enum import Enum
@@ -81,20 +82,21 @@ class LaneOffset:
   DUR_SLOW = 8.0 # [s] same, but slower for when position change is caused by auto lane offset
   STEP_SLOW = OFFSET / DUR_SLOW * DT_MDL * LANE_WIDTH_DEFAULT
   
-  AUTO_MIN_SHOULDER_WIDTH_FACTOR = 0.05 # [unitless] shoulder width must be as this this factor of lane width
+  AUTO_MIN_SHOULDER_WIDTH_FACTOR = 0.0 # [unitless] shoulder width must be as this this factor of lane width
   AUTO_MAX_LANE_WIDTH_FACTOR = 0.6 # [unitless] adjacent lanes whose width is narrower than this factor of lane width are considered a shoulder
   
   AUTO_MAX_PRED_LAT_ACCEL = 0.9 # [m/s^2] more than this predicted amount of lateral accel causes instant resuming of center position
   AUTO_MAX_CUR_LAT_ACCEL = 0.7 # same but for instantaneous vehicle lateral accel
   
-  AUTO_MIN_LANELINE_PROB = 0.3
+  AUTO_MIN_LANELINE_PROB = 0.75
   AUTO_MIN_ADJACENT_LANELINE_PROB = 0.15
   
-  AUTO_LANE_STATE_MIN_TIME = 4.0 # [s] amount of time the lane state must stay the same before it can be acted upon
+  AUTO_LANE_STATE_MIN_TIME = 3.0 # [s] amount of time the lane state must stay the same before it can be acted upon
   
   AUTO_ENABLE_ROAD_TYPES = {0, 10, 20, 30} # freeway and state highways (see highway ranks in /Users/haiiro/NoSync/optw/openpilot/selfdrive/mapd/lib/WayRelation.py)
   AUTO_ENABLE_ROAD_TYPE_MIN_SPEED = 40. * CV.MPH_TO_MS
   AUTO_ENABLE_TRAFFIC_MIN_SPEED = 10. * CV.MPH_TO_MS
+  AUTO_ENABLE_MIN_SPEED_DEADZONE = 2. # [m/s]
   
   AUTO_TRAFFIC_TIMEOUT = 30. # [s] amount of time auto lane position will be kept after last car is seen
   AUTO_TRAFFIC_MIN_TIME = 1.1 # [s] time an oncoming/ongoing car needs to be observed before it will be taken to indicate traffic
@@ -103,6 +105,8 @@ class LaneOffset:
   TRAFFIC_NEW_DETECT_STEER_CUTOFF = 9. # [degrees] instantaneous traffic isn't detected if steer angle greater than this
   AUTO_TRAFFIC_MIN_DIST_LANE_WIDTH_FACTOR = 1.25 # [unitless] factor of current lane width used to check against adjacent lead distance
   AUTO_TRAFFIC_STOPPED_MIN_DIST_LANE_WIDTH_FACTOR = 0.9 # [unitless] factor of current lane width used to check against adjacent lead distance, but for stopped objects
+  
+  ADJACENT_TRAFFIC_SEP_DIST_NONE = 2.5 # [s] separation follow distance used when <= 1 adjacent ongoing car
   
   def __init__(self, mass=0.):
     self.offset = 0.
@@ -136,8 +140,12 @@ class LaneOffset:
     self._right_traffic_last = LANE_TRAFFIC.NONE
     self._left_traffic_temp = LANE_TRAFFIC.NONE
     self._right_traffic_temp = LANE_TRAFFIC.NONE
-    self._right_traffic_last_seen_t = 0.
+    self._left_traffic_count = 0
+    self._right_traffic_count = 0
+    self._left_traffic_min_sep_dist = FirstOrderFilter(self.ADJACENT_TRAFFIC_SEP_DIST_NONE, 10.0, DT_MDL)
+    self._right_traffic_min_sep_dist = FirstOrderFilter(self.ADJACENT_TRAFFIC_SEP_DIST_NONE, 10.0, DT_MDL)
     self._left_traffic_last_seen_t = 0.
+    self._right_traffic_last_seen_t = 0.
     self._left_traffic_temp_t = 0.
     self._right_traffic_temp_t = 0.
     self._lprob_last = 0.
@@ -147,35 +155,39 @@ class LaneOffset:
   
   
   def do_auto_enable_traffic(self, ret, v_ego):
+    v_ego_diff = apply_deadzone(v_ego - self.AUTO_ENABLE_TRAFFIC_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
+    v_ego_diff_last = apply_deadzone(self._v_ego_last - self.AUTO_ENABLE_TRAFFIC_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
     if ret != AUTO_AUTO_LANE_MODE.ENGAGE and (self._left_traffic_last != self._left_traffic or self._right_traffic != self._right_traffic_last \
-      or (v_ego >= self.AUTO_ENABLE_TRAFFIC_MIN_SPEED and self._v_ego_last < self.AUTO_ENABLE_TRAFFIC_MIN_SPEED) \
-      or (v_ego < self.AUTO_ENABLE_TRAFFIC_MIN_SPEED and self._v_ego_last >= self.AUTO_ENABLE_TRAFFIC_MIN_SPEED) \
+      or (v_ego_diff > 0. and v_ego_diff_last <= 0.) \
+      or (v_ego_diff < 0. and v_ego_diff_last >= 0.) \
       or (self._lane_probs[1] >= self.AUTO_MIN_LANELINE_PROB and self._lprob_last <= self.AUTO_MIN_LANELINE_PROB) \
       or (self._lane_probs[1] < self.AUTO_MIN_LANELINE_PROB and self._lprob_last >= self.AUTO_MIN_LANELINE_PROB) \
       or (self._lane_probs[2] >= self.AUTO_MIN_LANELINE_PROB and self._rprob_last <= self.AUTO_MIN_LANELINE_PROB) \
       or (self._lane_probs[2] < self.AUTO_MIN_LANELINE_PROB and self._rprob_last >= self.AUTO_MIN_LANELINE_PROB)):  
-      if not self._auto_is_active and v_ego >= self.AUTO_ENABLE_TRAFFIC_MIN_SPEED \
+      if not self._auto_is_active and v_ego_diff > 0. \
         and (self._left_traffic != LANE_TRAFFIC.NONE or self._right_traffic != LANE_TRAFFIC.NONE) \
         and self._lane_probs[1] > self.AUTO_MIN_LANELINE_PROB and self._lane_probs[2] > self.AUTO_MIN_LANELINE_PROB:
           ret = AUTO_AUTO_LANE_MODE.ENGAGE
           self._auto_auto_enabled = True
       elif self._auto_is_active and self._auto_auto_enabled \
-        and (self._left_traffic == LANE_TRAFFIC.NONE and self._right_traffic == LANE_TRAFFIC.NONE \
-          or v_ego < self.AUTO_ENABLE_TRAFFIC_MIN_SPEED \
-          or (self._lane_probs[1] < self.AUTO_MIN_LANELINE_PROB and self._lane_probs[2] < self.AUTO_MIN_LANELINE_PROB)):
+        and (self._left_traffic == LANE_TRAFFIC.NONE and self._right_traffic == LANE_TRAFFIC.NONE) \
+          or v_ego_diff < 0. \
+          or (self._lane_probs[1] < self.AUTO_MIN_LANELINE_PROB and self._lane_probs[2] < self.AUTO_MIN_LANELINE_PROB):
             ret = AUTO_AUTO_LANE_MODE.DISENGAGE
             self._auto_auto_enabled = False
     return ret
 
   def do_auto_enable_lanelines(self, ret, road_type, v_ego):
+    v_ego_diff = apply_deadzone(v_ego - self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
+    v_ego_diff_last = apply_deadzone(self._v_ego_last - self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED, self.AUTO_ENABLE_MIN_SPEED_DEADZONE)
     if road_type != self._road_type_last or ret == AUTO_AUTO_LANE_MODE.DISENGAGE \
-        or (v_ego >= self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED and self._v_ego_last < self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED) \
-        or (v_ego < self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED and self._v_ego_last >= self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED):
-      if (road_type in self.AUTO_ENABLE_ROAD_TYPES and v_ego >= self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED and (not self._auto_is_active or ret == AUTO_AUTO_LANE_MODE.DISENGAGE)):
+        or (v_ego_diff > 0. and v_ego_diff_last <= 0.) \
+        or (v_ego_diff < 0. and v_ego_diff_last >= 0.):
+      if (road_type in self.AUTO_ENABLE_ROAD_TYPES and v_ego_diff > 0. and (not self._auto_is_active or ret == AUTO_AUTO_LANE_MODE.DISENGAGE)):
         ret = AUTO_AUTO_LANE_MODE.NO_CHANGE if ret == AUTO_AUTO_LANE_MODE.DISENGAGE else AUTO_AUTO_LANE_MODE.ENGAGE
         self._auto_auto_enabled = True
       elif self._auto_is_active and self._auto_auto_enabled \
-        and (road_type not in self.AUTO_ENABLE_ROAD_TYPES or v_ego < self.AUTO_ENABLE_ROAD_TYPE_MIN_SPEED) :
+        and (road_type not in self.AUTO_ENABLE_ROAD_TYPES or v_ego_diff < 0.) :
           ret = AUTO_AUTO_LANE_MODE.DISENGAGE
           self._auto_auto_enabled = False
     return ret
@@ -213,28 +225,19 @@ class LaneOffset:
       
       # get road edge "probabilities"
       self._road_edge_probs = [ interp(md.roadEdgeStds[i], [.3, 1.], [1.0, 0.0]) for i in range(2) ]
+      self._lane_probs = md.laneLineProbs
             
       # get laneline probabilities
-      for i in range(3): 
-        l_prob, r_prob = md.laneLineProbs[i], md.laneLineProbs[i+1]
-        if ((i == 0 and l_prob > self.AUTO_MIN_ADJACENT_LANELINE_PROB) or l_prob > self.AUTO_MIN_LANELINE_PROB) \
-            and ((i == 2 and r_prob > self.AUTO_MIN_ADJACENT_LANELINE_PROB) or r_prob > self.AUTO_MIN_LANELINE_PROB):
-          # Reduce reliance on uncertain lanelines (less stringent for adjacent lanelines)
-          rll_y, lll_y = np.array(md.laneLines[i+1].y), np.array(md.laneLines[i].y)
-          
-          if i == 0: # set left adjacent 
-            self._lane_probs[0] = l_prob
-            self._lane_width_mean_left_adjacent = np.mean(rll_y - lll_y)
-            if self._road_edge_probs[0] > 0.:
-              self._shoulder_width_mean_left = np.mean(lll_y - np.array(md.roadEdges[0].y))
-          elif i == 1: # set center
-            self._lane_probs[1] = l_prob
-            self._lane_probs[2] = r_prob
-          else: # set right 
-            self._lane_probs[3] = r_prob
-            self._lane_width_mean_right_adjacent = np.mean(rll_y - lll_y)
-            if self._road_edge_probs[1] > 0.:
-              self._shoulder_width_mean_right = np.mean(np.array(md.roadEdges[1].y) - rll_y)
+      for i in [0,2]: 
+        rll_y, lll_y = np.array(md.laneLines[i+1].y), np.array(md.laneLines[i].y)
+        if i == 0: # set left adjacent 
+          self._lane_width_mean_left_adjacent = np.mean(rll_y - lll_y)
+          if self._road_edge_probs[0] > 0.:
+            self._shoulder_width_mean_left = np.mean(lll_y - np.array(md.roadEdges[0].y))
+        elif i == 2: # set right 
+          self._lane_width_mean_right_adjacent = np.mean(rll_y - lll_y)
+          if self._road_edge_probs[1] > 0.:
+            self._shoulder_width_mean_right = np.mean(np.array(md.roadEdges[1].y) - rll_y)
 
       
       # see if adjacent lane can be treated as shoulder because it's too narrow
@@ -252,22 +255,29 @@ class LaneOffset:
     if self._cs is None or abs(self._cs.steeringAngleDeg) > self.TRAFFIC_NEW_DETECT_STEER_CUTOFF:
       return
     
+    sep_dist = self.ADJACENT_TRAFFIC_SEP_DIST_NONE # [s] if <= 1 adjacent cars, this is the "separation" between them
+    
     leads = rs.leadsLeft
     if len(leads) > 0:
       check_lane_width = lane_width * self.AUTO_TRAFFIC_MIN_DIST_LANE_WIDTH_FACTOR
       if md.laneLineProbs[0] >= self.AUTO_MIN_ADJACENT_LANELINE_PROB and md.laneLineProbs[1] >= self.AUTO_MIN_LANELINE_PROB \
         and md.laneLines[1].y[0] > -2.2:
-          lv = [l.vLeadK for l in leads if lead_between_lines(md.laneLines[0], md.laneLines[1], l)]
+          l1 = [l for l in leads if lead_between_lines(md.laneLines[0], md.laneLines[1], l)]
       elif md.laneLineProbs[1] >= self.AUTO_MIN_LANELINE_PROB and md.laneLines[1].y[0] > -2.:
-        lv = [l.vLeadK for l in leads if lead_close_to_line(md.laneLines[1], l, lane_width, True)]
+        l1 = [l for l in leads if lead_close_to_line(md.laneLines[1], l, lane_width, True)]
       else:
-        lv = [l.vLeadK for l in leads if abs(l.dPath) < check_lane_width]
+        l1 = [l for l in leads if abs(l.dPath) < check_lane_width]
         
-      if len(lv) > 0:
+      self._left_traffic_count = len(l1)
+      if len(l1) > 0:
+        lv = [l.vLeadK for l in l1]
         min_v, max_v, mean_v = min_max_mean(lv)
         check_v = min_v if abs(min_v - mean_v) < abs(max_v - mean_v) else max_v
         if check_v > self.AUTO_TRAFFIC_MIN_SPEED and self._lane_width_mean_left_adjacent > 0.:
           left_traffic = LANE_TRAFFIC.ONGOING
+          if len(l1) > 1:
+            l1.sort(key= lambda x:x.dRel)
+            sep_dist = min([(ldj.dRel - ldi.dRel) / (abs(ldi.vLeadK) + 0.1) for ldi,ldj in zip(l1[:-1], l1[1:])])
         elif check_v < -self.AUTO_TRAFFIC_MIN_SPEED:
           left_traffic = LANE_TRAFFIC.ONCOMING
         else:
@@ -275,23 +285,30 @@ class LaneOffset:
           lv = [l.vLeadK for l in rs.leadsLeft if abs(l.dPath) < check_lane_width]
           if len(lv) > 0:
             left_traffic = LANE_TRAFFIC.STOPPED
-    
+    self._left_traffic_min_sep_dist.update(min(sep_dist, self.ADJACENT_TRAFFIC_SEP_DIST_NONE))
+          
+    sep_dist = self.ADJACENT_TRAFFIC_SEP_DIST_NONE # [s] if <= 1 adjacent cars, this is the "separation" between them
     leads = rs.leadsRight
     if len(leads) > 0:
       check_lane_width = lane_width * self.AUTO_TRAFFIC_MIN_DIST_LANE_WIDTH_FACTOR
       if md.laneLineProbs[3] >= self.AUTO_MIN_ADJACENT_LANELINE_PROB and md.laneLineProbs[2] >= self.AUTO_MIN_LANELINE_PROB \
         and md.laneLines[2].y[0] < 2.2:
-          lv = [l.vLeadK for l in leads if lead_between_lines(md.laneLines[2], md.laneLines[3], l)]
+          l1 = [l for l in leads if lead_between_lines(md.laneLines[2], md.laneLines[3], l)]
       elif md.laneLineProbs[2] >= self.AUTO_MIN_LANELINE_PROB and md.laneLines[2].y[0] < 2.:
-        lv = [l.vLeadK for l in leads if lead_close_to_line(md.laneLines[2], l, lane_width, False)]
+        l1 = [l for l in leads if lead_close_to_line(md.laneLines[2], l, lane_width, True)]
       else:
-        lv = [l.vLeadK for l in leads if abs(l.dPath) < check_lane_width]
+        l1 = [l for l in leads if abs(l.dPath) < check_lane_width]
       
-      if len(lv) > 0:
+      self._right_traffic_count = len(l1)
+      if len(l1) > 0:
+        lv = [l.vLeadK for l in l1]
         min_v, max_v, mean_v = min_max_mean(lv)
         check_v = min_v if abs(min_v - mean_v) < abs(max_v - mean_v) else max_v
         if check_v > self.AUTO_TRAFFIC_MIN_SPEED and self._lane_width_mean_right_adjacent > 0.:
           right_traffic = LANE_TRAFFIC.ONGOING
+          if len(l1) > 1:
+            l1.sort(key= lambda x:x.dRel)
+            sep_dist = min([(ldj.dRel - ldi.dRel) / (abs(ldi.vLeadK) + 0.1) for ldi,ldj in zip(l1[:-1], l1[1:])])
         elif check_v < -self.AUTO_TRAFFIC_MIN_SPEED:
           right_traffic = LANE_TRAFFIC.ONCOMING
         else:
@@ -299,6 +316,7 @@ class LaneOffset:
           lv = [l.vLeadK for l in rs.leadsRight if abs(l.dPath) < check_lane_width]
           if len(lv) > 0:
             right_traffic = LANE_TRAFFIC.STOPPED
+    self._right_traffic_min_sep_dist.update(min(sep_dist, self.ADJACENT_TRAFFIC_SEP_DIST_NONE))
     
     if left_traffic != LANE_TRAFFIC.NONE:
       if self._left_traffic_temp != left_traffic:
@@ -329,11 +347,11 @@ class LaneOffset:
     timeout_override = False
     
     if not (self._cs is None or self._cs.leftBlinker or self._cs.rightBlinker):
-      if self._left_traffic != LANE_TRAFFIC.NONE \
+      if self._left_traffic in [LANE_TRAFFIC.ONCOMING, LANE_TRAFFIC.ONGOING] \
           and self._right_traffic == LANE_TRAFFIC.NONE:
         lane_pos_auto = -1.
         timeout_override = True
-      elif self._right_traffic != LANE_TRAFFIC.NONE \
+      elif self._right_traffic in [LANE_TRAFFIC.ONCOMING, LANE_TRAFFIC.ONGOING] \
           and self._left_traffic == LANE_TRAFFIC.NONE:
         lane_pos_auto = 1.
         timeout_override = True